orchestra-ai/pyproject.toml
[project]
name = "orchestra-ai"
version = "0.1.0"
description = "A modular agent orchestration framework (CLI + API) with tools, memory, and routing."
requires-python = ">=3.10"
dependencies = [
  "openai>=1.30.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
orchestra = "orchestra.cli:app"

[tool.ruff]
line-length = 100
